Enhance BuildPlan
Previously, during detect any existing keys that needed to be contributed were
overwritten with an empty Dependency declaration.  This is a poor behavior as
the only requirement was that a key exist, not the contents of the Dependency.
This change attempts to retrieve an existing Dependency for a key which will
return an empty declaration if it does not yet exist, resulting in the same
effect for unspecified keys, and passes through the previous key if it
existed.

Signed-off-by: Ben Hale <bhale@pivotal.io>